Move from model namespace
