+++ src/Model/ApplicationResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Model;
 
 use DateTimeInterface;
