+++ src/Model/ErrorCodeList.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Model;
 
 final class ErrorCodeList
 {
