+++ src/Model/GetStatusCdrResponse.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Model;
 
 class GetStatusCdrResponse
 {
     /**
      * @var StatusCdrResponse
      */
-    private $statusCdr;
+    public $statusCdr;
 }
