+++ src/Model/GetStatusRequest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Model;
 
-class SendPackResponse
+class GetStatusRequest
 {
     /**
      * @var string
